+++ common/src/main/java/me/drex/antixray/util/ChunkPacketBlockControllerAntiXray.java
@@ -7,7 +7,6 @@ import me.drex.antixray.interfaces.IChunkPacket;
 import me.drex.antixray.interfaces.ILevel;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
-import net.minecraft.core.Registry;
 import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.core.registries.Registries;
 import net.minecraft.network.protocol.game.ClientboundLevelChunkWithLightPacket;