+++ CHANGELOG.md
@@ -5,11 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-### Fixed
-- Typo in forge release names
 
+## [1.2.7] - 2022-12-14
 ### Added
 - Bypass permission: "antixray.bypass"
+
+### Fixed
+- Typo in forge release names
 
 ## [1.2.6] - 2022-09-08
 ### Added